backend/requirements.txt
flask
flask-cors
flask-jwt-extended
pymupdf
openai
requests
python-dotenv
werkzeug
